Django/Lecture #22 - Django ORM lection/requirements.txt
amqp==5.1.1
anyio==3.7.0
asgiref==3.6.0
async-timeout==4.0.2
billiard==4.1.0
celery==5.3.0
certifi==2023.5.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.1
django-dotenv==1.4.2
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
kombu==5.3.1
mysqlclient==2.1.1
Pillow==9.5.0
prompt-toolkit==3.0.38
PyJWT==2.7.0
python-dateutil==2.8.2
python-telegram-bot==20.3
pytz==2023.3
redis==4.5.5
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
zope.event==4.6
zope.interface==6.0
